LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-range-normal-matches-cover-expected.txt


FAIL Changing the animation range updates the play state assert_equals: expected (string) "normal" but got (object) object "[object Object]"
